Add DijkstraMap, wow.
The only thing currently missing is the Technique-based pathfinding, because I haven't brought over Technique yet (and it might need reworking anyway).

+++ squidgrid/src/main/java/com/github/yellowstonegames/grid/Measurement.java
@@ -27,11 +27,11 @@ public enum Measurement {
 	 */
 	EUCLIDEAN;
 
-	public double heuristic(Direction target) {
+	public float heuristic(Direction target) {
 		if (this == Measurement.EUCLIDEAN && (target == Direction.DOWN_LEFT || target == Direction.DOWN_RIGHT || target == Direction.UP_LEFT || target == Direction.UP_RIGHT)) {
-			return 1.4142135623730951; //Math.sqrt(2.0);
+			return 1.4142135623730951f; //Math.sqrt(2.0);
 		}
-		return 1.0;
+		return 1f;
 	}
 
 	public int directionCount() {
